move constants from doc.go to const.go files
