+++ feg/gateway/services/radiusd/const.go
@@ -11,4 +11,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ctraced
+package radiusd
+
+const (
+	ServiceName = "radiusd"
+)
