+++ orc8r/cloud/go/services/configurator/const.go
@@ -1,0 +1,25 @@
+/*
+* Copyright 2020 The Magma Authors.
+*
+* This source code is licensed under the BSD-style license found in the
+* LICENSE file in the root directory of this source tree.
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+ */
+
+package configurator
+
+const (
+	// ServiceName is the name of this service.
+	ServiceName = "configurator"
+
+	// NetworkConfigSerdeDomain is the Serde domain for network configs.
+	NetworkConfigSerdeDomain = "configurator_network_configs"
+
+	// NetworkEntitySerdeDomain is the Serde domain for network entity configs.
+	NetworkEntitySerdeDomain = "configurator_entity_configs"
+)
