+++ orc8r/cloud/go/services/configurator/doc.go
@@ -69,14 +69,3 @@
 	provide an mconfig builder servicer.
 */
 package configurator
-
-const (
-	// ServiceName is the name of this service.
-	ServiceName = "CONFIGURATOR"
-
-	// NetworkConfigSerdeDomain is the Serde domain for network configs.
-	NetworkConfigSerdeDomain = "configurator_network_configs"
-
-	// NetworkEntitySerdeDomain is the Serde domain for network entity configs.
-	NetworkEntitySerdeDomain = "configurator_entity_configs"
-)
