+++ orc8r/cloud/go/services/eventd/const.go
@@ -11,4 +11,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ctraced
+package eventd
+
+const ServiceName = "eventd"
