+++ orc8r/cloud/go/services/eventd/doc.go
@@ -12,5 +12,3 @@ limitations under the License.
 */
 
 package eventd
-
-const ServiceName = "EVENTD"
