+++ orc8r/cloud/go/services/metricsd/const.go
@@ -22,4 +22,5 @@ const (
 	AlertmanagerApiURL           = "alertmanagerApiURL"
 
 	UseSeriesCache = "useSeriesCache"
+	ServiceName    = "metricsd"
 )
