+++ orc8r/cloud/go/services/state/const.go
@@ -14,3 +14,9 @@
 // Package state contains the state service, an orchestrator microservice which
 // manages device-reported state.
 package state
+
+const (
+	ServiceName = "state"
+	SerdeDomain = "state"
+	DBTableName = "states"
+)
